game_bot/requirements.txt
amplitude-analytics==1.1.4
humanize==4.8.0
Pillow==9.1.0
py-cord~=2.4
python-dotenv==0.19.2
pytest-asyncio~=0.21
pytest-bdd~=6.1
pytest-cov~=3.0
pytest-django~=4.5
pytest-factoryboy~=2.1
pytest-lazy-fixture~=0.6
pytest-mock~=3.7
pytz~=2023.3
requests==2.26.0
sentry-sdk==1.30.0
unittest-xml-reporting==3.0.4
